% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddDist}
\alias{AddDist}
\title{Generates a distance matrix for the active assay of the given object.}
\usage{
AddDist(pisces.obj, use.pcs = TRUE, dist.method = c("cor", "l1", "l2"))
}
\arguments{
\item{pisces.obj}{A Seurat object w/ either 'scale.data' in the active assay or 'pca' object in the 'misc' field of the active assay.}

\item{use.pcs}{Flag to indicate whether PCA should be used. If true, object must have 'pca' object in the 'misc' field of the active assay.}

\item{dist.method}{One of 'cor', 'l1', or 'l2'. Default of 'cor'.}
}
\value{
Seurat object with a 'dist' object added to the 'misc' field of the active assay.
}
\description{
Generates a distance matrix for the active assay of the given object.
}
