% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddGES}
\alias{AddGES}
\title{Generates a GES, either internally or against an external reference.}
\usage{
AddGES(pisces.obj, method = c("scale", "ecdf"), ref.mat = NULL)
}
\arguments{
\item{pisces.obj}{A Seurat object with the `PISCESpact` assay.}

\item{method}{One of `scale` or `ecdf`. If not specified, will use `scale`.}

\item{ref.mat}{Optional matrix of appropriately normalized reference data (features X samples).}
}
\value{
A Seurat object with a `ges` matrix (features X samples) added to the `PISCESpact` assay.
}
\description{
Generates a GES, either internally or against an external reference.
}
