% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddMDS}
\alias{AddMDS}
\title{Generates a Multidimensional Scaling (MDS) for the active assay of the given object.
Checks for a distance object in the 'misc' field; calculates one based on correlation distance if not present.}
\usage{
AddMDS(pisces.obj, num.dims = 2)
}
\arguments{
\item{pisces.obj}{A Seurat object with 'scale.data' in the active assay.}

\item{num.dims}{Number of dimensions to return. Default of 2.}
}
\value{
A Seurat object with a 'mds' object added to the 'misc' field of the active assay.
}
\description{
Generates a Multidimensional Scaling (MDS) for the active assay of the given object.
Checks for a distance object in the 'misc' field; calculates one based on correlation distance if not present.
}
