% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddNetworks}
\alias{AddNetworks}
\title{Adds networks to a Seurat object with a `PISCESpact` assay.}
\usage{
AddNetworks(pisces.obj, net.list)
}
\arguments{
\item{pisces.obj}{Seurat object with `PISCESpact`}

\item{net.list}{List of ARACNe3 networks.}
}
\value{
Seurat object with `a3.nets` object added to `misc` field of `PISCESpact` assay.
}
\description{
Adds networks to a Seurat object with a `PISCESpact` assay.
}
