% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddPCA}
\alias{AddPCA}
\title{Generates a principle component analysis (PCA) for the active assay of the given object.}
\usage{
AddPCA(pisces.obj, num.pcs = 100)
}
\arguments{
\item{pisces.obj}{A Seurat object with 'scale.data' in the active assay.}

\item{num.pcs}{Flag to use principle components. Default of 100.}
}
\value{
A Seurat object with a 'pca' object added to the 'misc' field of the active assay.
}
\description{
Generates a principle component analysis (PCA) for the active assay of the given object.
}
