% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddPISCESAssay}
\alias{AddPISCESAssay}
\title{Adds a named assay 'PISCES' to a Seurat object.}
\usage{
AddPISCESAssay(seurat.obj, stage)
}
\arguments{
\item{seurat.obj}{Seurat object w/ RNA assay.}

\item{stage}{One of 'gexp' or 'pact'; adds either the 'PISCES.gexp' or 'PISCES.pact' assay. Default of 'pact'.}
}
\value{
Seurat object w/ original assay and added 'PISCES' assay.
}
\description{
Adds a named assay 'PISCES' to a Seurat object.
}
