% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{AddUMAP}
\alias{AddUMAP}
\title{Generates a UMAP for the active assay of the given object.}
\usage{
AddUMAP(pisces.obj, dense.map = FALSE)
}
\arguments{
\item{pisces.obj}{A Seurat object with 'scale.data' in the active assay.}

\item{dense.map}{Flag to return dense-UMAP. Default of FALSE.}
}
\value{
A Seurat object with a 'umap' object added to the 'misc' field of the active assay.
}
\description{
Generates a UMAP for the active assay of the given object.
}
