% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{AnovaMRs}
\alias{AnovaMRs}
\title{Identifies MRs based on ANOVA analysis for a given clustering.}
\usage{
AnovaMRs(dat.mat, clustering)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{clustering}{Clustering vector}
}
\value{
A named vector of p-values for each protein
}
\description{
Identifies MRs based on ANOVA analysis for a given clustering.
}
