% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne_funcs.R
\name{ARACNeTable}
\alias{ARACNeTable}
\title{Saves a matrix in a format for input to ARACNe}
\usage{
ARACNeTable(dat.mat, out.file, subset = TRUE)
}
\arguments{
\item{dat.mat}{Matrix of data (genes X samples).}

\item{out.file}{Output file where matrix will be saved.}

\item{subset}{Switch for subsetting the matrix to 500 samples. Default TRUE.}
}
\description{
Saves a matrix in a format for input to ARACNe
}
