% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{BTTestMRs}
\alias{BTTestMRs}
\title{Identifies MRs based on a bootstraped Ttest between clusters.}
\usage{
BTTestMRs(dat.mat, class.vect, bootstrap.num = 100)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{class.vect}{Vector of cluster labels / class vectors.}

\item{bootstrap.num}{Number of bootstraps to use. Default of 10}
}
\value{
Returns a list of lists; each list is a vector of sorted log p-values for each cluster.
}
\description{
Identifies MRs based on a bootstraped Ttest between clusters.
}
