% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-reduc_funcs.R
\name{cbcMR_UMAP}
\alias{cbcMR_UMAP}
\title{Builds a UMAP transformation based on the shared, cell-by-cell master regulators. Utilizes custom UMAP parameters ('pearson', 25 neighbors).}
\usage{
cbcMR_UMAP(dat.mat, num.mrs = 50)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (features X samples).}

\item{num.mrs}{Number of top master regulators to take from each cell. Default of 50.}
}
\value{
Returns a UMAP based on the cell-by-cell master regulators.
}
\description{
Builds a UMAP transformation based on the shared, cell-by-cell master regulators. Utilizes custom UMAP parameters ('pearson', 25 neighbors).
}
