% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{CBCMRs}
\alias{CBCMRs}
\title{Identifies MRs on a cell-by-cell basis and returns a merged, unique list of all such MRs.}
\usage{
CBCMRs(dat.mat, numMRs = 25)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{numMRs}{Default number of MRs to identify in each cell. Default of 25.}
}
\value{
Returns a list of master regulators, the unique, merged set from all cells.
}
\description{
Identifies MRs on a cell-by-cell basis and returns a merged, unique list of all such MRs.
}
