% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{cluster_mr_volcano}
\alias{cluster_mr_volcano}
\title{Generates cluster-specific volcano plots from the given master regulator object.}
\usage{
cluster_mr_volcano(mr.obj, num.mrs = 10)
}
\arguments{
\item{mr.obj}{Master regulator object, such as that generated by `cluster_signature_mrs`.
Should be a list with a `cluster.narnea` element, containing cluster-specific NaRnEA results (NES and PES matrices).}

\item{num.mrs}{Number of top / bottom MRs to label. Default of 10.}
}
\value{
A ggplot object.
}
\description{
Generates cluster-specific volcano plots from the given master regulator object.
}
