% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{cluster_pair_mr_volcano}
\alias{cluster_pair_mr_volcano}
\title{Generates cluster-specific volcano plots from the given master regulator object.}
\usage{
cluster_pair_mr_volcano(mr.obj, clust.vec, num.mrs = 10)
}
\arguments{
\item{mr.obj}{Master Regulator object generated by pairwise_cluster_mrs.
Should be a list with a `mr.narnea` element, containing cluster-v-cluster NaRnEA results (NES and PES matrices).
Names should be structured as `cn1.v.cn2`}

\item{clust.vec}{Clustering vector.}

\item{num.mrs}{Number of top / bottom MRs to label. Default of 10.}
}
\value{
A ggplot object.
}
\description{
Generates cluster-specific volcano plots from the given master regulator object.
}
