% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{cluster_signature_mrs}
\alias{cluster_signature_mrs}
\title{Identifies cluster specific master regulators by generating cluter-specific signatures and running NaRnEA.}
\usage{
cluster_signature_mrs(norm.counts, clust.vec, net.list, num.mrs = 10)
}
\arguments{
\item{norm.counts}{Matrix of normalized counts (features X samples).}

\item{clust.vec}{Clustering vector.}

\item{net.list}{List of ARACNe3 networks.}

\item{num.mrs}{Number of master regulators to return for each cluster. Default of 10.}
}
\value{
A list; table of positive MRs (`positive`) and negative MRs (`negative`); 
the cluster-specific NaRnEA results (`cluster.narnea`).
}
\description{
Identifies cluster specific master regulators by generating cluter-specific signatures and running NaRnEA.
}
