% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{ClusterColors}
\alias{ClusterColors}
\title{Identifies a vector of colors for a given number of clusters. Internal function.}
\usage{
ClusterColors(k, offset = 0)
}
\arguments{
\item{k}{Number of clusters.}

\item{offset}{Optional argument to shift colors along color wheel.}
}
\value{
A vector of hues
}
\description{
Identifies a vector of colors for a given number of clusters. Internal function.
}
