% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{cohen_kappa}
\alias{cohen_kappa}
\title{Calculates the Cohen's Kappa between two categorical vectors.}
\usage{
cohen_kappa(vec.a, vec.b)
}
\arguments{
\item{vec.a}{First vector of categorical labels.}

\item{vec.b}{Second vector of categorical labels. Assumes same ordering as `vec.a`.}
}
\value{
Cohen's Kappa
}
\description{
Calculates the Cohen's Kappa between two categorical vectors.
}
