% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{color_levels}
\alias{color_levels}
\title{Returns color gradient for the specified data type (green/purple for Gene Expression; red/blue for proteina ctivity)}
\usage{
color_levels(data.type = c("gexp", "pact"), num.colors)
}
\arguments{
\item{data.type}{Type of data to use; either 'gexp' or 'pact'}

\item{num.colors}{Number of colors to return.}
}
\value{
Vector of colors.
}
\description{
Returns color gradient for the specified data type (green/purple for Gene Expression; red/blue for proteina ctivity)
}
