% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{ColorLevels}
\alias{ColorLevels}
\title{Returns color gradient for the specified data type (green/purple for Gene Expression; red/blue for VIPER)}
\usage{
ColorLevels(num.colors, data.type)
}
\arguments{
\item{data.type}{Type of data to use; either 'gexp' or 'vip'}

\item{num.breaks}{Number of breaks in the gradient.}
}
\value{
Vector of colors.
}
\description{
Returns color gradient for the specified data type (green/purple for Gene Expression; red/blue for VIPER)
}
