% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{combine_nes}
\alias{combine_nes}
\title{Generates a final NES from the combination of the Directed and Undirected NES.}
\usage{
combine_nes(D.nes, U.nes, COV.nes)
}
\arguments{
\item{D.nes}{NES scores for the directed enrichment (regulons x samples).}

\item{U.nes}{NES scores for the undirected enrichment (regulons x samples).}

\item{COV.nes}{Covariance of the undirected and directed enrichment NES (regulons x samples).}
}
\value{
Matrix of integrated NES scores (regulons x samples).
}
\description{
Generates a final NES from the combination of the Directed and Undirected NES.
}
