% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funcs.R
\name{convert_genes}
\alias{convert_genes}
\title{Converts a list of gene names from 'from' to 'to' conventions.}
\usage{
convert_genes(gene.list, species, from, to)
}
\arguments{
\item{gene.list}{List of genes to be converted.}

\item{species}{Species for this data. Either 'human' or 'mouse'.}

\item{from}{Starting name convention. One of 'ensg', 'gn', or 'entrez'.}

\item{to}{Ending name convention. One of 'ensg', 'gn', or 'entrez'.}
}
\value{
Matrix with converted row names.
}
\description{
Converts a list of gene names from 'from' to 'to' conventions.
}
