% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_funcs.R
\name{cor_dist}
\alias{cor_dist}
\title{Generates a distance matrix using the square root of (1 - spearman).}
\usage{
cor_dist(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of data (features X samples).}
}
\value{
Distance structure.
}
\description{
Generates a distance matrix using the square root of (1 - spearman).
}
