% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{CorDist}
\alias{CorDist}
\title{Generates a distance matrix using sqrt(1-cor(x)) as a distance metric.
If a seurat object, stores results in cor.dist (misc); otherwise, returns distance matrix.}
\usage{
CorDist(data.object, use.scaled = FALSE, cor.method = "spearman")
}
\arguments{
\item{data.object}{Either a Seurat object or a matrix of data (features X samples).}

\item{use.scaled}{If specified AND if data.object is a Seurat object, will use the "scale.data" field instead of "data"}

\item{cor.method}{Method argument for cor function; spearman by default.}
}
\value{
Rank transformed matrix, or appropriately adjusted seurat object.
}
\description{
Generates a distance matrix using sqrt(1-cor(x)) as a distance metric.
If a seurat object, stores results in cor.dist (misc); otherwise, returns distance matrix.
}
