% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{cpm_norm}
\alias{cpm_norm}
\title{Peforms a CPM normalization on the given matrix.}
\usage{
cpm_norm(dat.mat, l2 = FALSE, remove.zeroes = TRUE)
}
\arguments{
\item{dat.mat}{Matrix of data, typically unnormalized counts (genes X samples).}

\item{l2}{If true, will log transform the matrix. Default of FALSE.}

\item{remove.zeroes}{Removes rows (genes) with zero expression across all samples. Default of TRUE.}
}
\value{
CPM normalizaed matrix.
}
\description{
Peforms a CPM normalization on the given matrix.
}
