% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-reduc_funcs.R
\name{CustomUMAP}
\alias{CustomUMAP}
\title{Generates a UMAP based on a set of proteins.}
\usage{
CustomUMAP(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (features X samples).}

\item{mrs}{List of proteins to use as master regulators.}
}
\value{
UMAP object.
}
\description{
Generates a UMAP based on a set of proteins.
}
