% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{directed_nes}
\alias{directed_nes}
\title{Computes the Directed NES}
\usage{
directed_nes(R, S, AW.AM.prod, E.rs, E.r2, n)
}
\arguments{
\item{R}{Rank normalized ges matrix (genes x samples).}

\item{S}{Sign of ges matrix (genes x samples).}

\item{AW.AM.prod}{Product of the association weight (AW) and mode (AM) matrices created from the regulon list.}

\item{E.rs}{Expected value of r*s in each sample (1 x samples).}

\item{E.r2}{Expected value of r^2 in each sample (scalar value).}

\item{n}{Number of samples (scalar value).}
}
\value{
List of matrices, each (regulons x samples):
Enrichment score matrix 'es'; expected value 'exp'; variance 'var'; NES matrix 'NES'
}
\description{
Computes the Directed NES
}
