% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{ecdf_ges}
\alias{ecdf_ges}
\title{Generates a gene expression signature through an ECDF, either internally or against a reference.}
\usage{
ecdf_ges(dat.mat, ref.mat)
}
\arguments{
\item{dat.mat}{Matrix of test data (features X samples).}

\item{ref.mat}{Optional matrix of reference data (features X samples). If not specified, an internal GES will be generated.}
}
\value{
GES Matrix (featues X samples).
}
\description{
Generates a gene expression signature through an ECDF, either internally or against a reference.
}
