% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{ecdf_norm}
\alias{ecdf_norm}
\title{Normalizes values in the test vector based on the ECDF of the reference vector.}
\usage{
ecdf_norm(test.vec, ref.vec)
}
\arguments{
\item{test.vec}{Vector of test data.}

\item{ref.vec}{Vector of reference data.}
}
\value{
Vector of normalized values as z-scores.
}
\description{
Normalizes values in the test vector based on the ECDF of the reference vector.
}
