% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{ecdfGES}
\alias{ecdfGES}
\title{Generates a Gene Expression Signature using an ECDF of the reference.}
\usage{
ecdfGES(data.object, ref.mat = NULL)
}
\arguments{
\item{ref.mat}{Optional reference matrix. If not included, an internal signature is performed.}

\item{test.mat}{Matrix of normalized gene expression, i.e. CPM (features X samples)}
}
\value{
GES matrix, or appropriately adjusted seurat object.
}
\description{
Generates a Gene Expression Signature using an ECDF of the reference.
}
