% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-name_funcs.R
\name{Ensemble2GeneName}
\alias{Ensemble2GeneName}
\title{Transforms gene names from Ensembl to hgnc.}
\usage{
Ensemble2GeneName(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of data with ENSEMBL names (genes X samples).}
}
\value{
Data with HGNC names. Some data will likely be lost in the conversion.
}
\description{
Transforms gene names from Ensembl to hgnc.
}
