% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-name_funcs.R
\name{Entrez2Ensemble}
\alias{Entrez2Ensemble}
\title{Transforms gene names from Entrez to Ensembl}
\usage{
Entrez2Ensemble(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of data with Entrez names (genes X samples).}
}
\value{
Data with Ensembl names. Some data will likely be lost in the conversion.
}
\description{
Transforms gene names from Entrez to Ensembl
}
