% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_funcs.R
\name{fast_pca}
\alias{fast_pca}
\title{Generates the specified number of PCs using RSpectra.}
\usage{
fast_pca(dat.mat, num.pcs = 10)
}
\arguments{
\item{dat.mat}{Matrix of data (features X samples).}

\item{num.pcs}{Number of PCs to be generated. Default of 10.}
}
\value{
A list, with with pca matrix (x), a vector of feature sds (sdev), and the rotation matrix (rotation).
}
\description{
Generates the specified number of PCs using RSpectra.
}
