% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{GESTransform}
\alias{GESTransform}
\title{Generates a gene expression signature (GES) using internal normalization.
If a seurat object, stores results in scale.data field; otherwise, returns GES matrix.}
\usage{
GESTransform(data.object)
}
\arguments{
\item{data.object}{Either a Seurat object or a matrix of raw count data (genes X samples).}
}
\value{
GES matrix, or appropriately adjusted seurat object.
}
\description{
Generates a gene expression signature (GES) using internal normalization.
If a seurat object, stores results in scale.data field; otherwise, returns GES matrix.
}
