% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{get_mr_set}
\alias{get_mr_set}
\title{Returns set of candidate master regulators or markers from the given mr.list object.}
\usage{
get_mr_set(
  mr.list,
  num.mrs = 10,
  reg.class = c("regulator", "marker"),
  reg.sign = c("pos", "neg")
)
}
\arguments{
\item{mr.list}{MR List, such as that generated by `stouffer_cluster_mrs` or `kappa_cluster_mrs`.}

\item{num.mrs}{Number of MRs to return for each cluster. Defualt of 10.}

\item{reg.class}{Class of proteins to return. One of `c('regulator', 'marker')`, with `regulator` as default.}

\item{reg.sign}{Sign of proteins to return. One of `c('pos', 'neg')`, with `pos` as default.}
}
\value{
A dataframe with proteins and assocaited clusters.
}
\description{
Returns set of candidate master regulators or markers from the given mr.list object.
}
