% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{GetMRs}
\alias{GetMRs}
\title{Returns the master regulators for the given data.}
\usage{
GetMRs(dat.mat, clustering, method, numMRs = 50, bottom = FALSE, weights, ...)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{clustering}{Optional argument for a vector of cluster labels.}

\item{method}{'Stouffer' or 'ANOVA'}

\item{numMRs}{Number of MRs to return per cluster. Default of 50.}

\item{bottom}{Switch to return downregulated proteins in MR list. Default FALSE>}

\item{weights}{Optional argument for weights, which can be used in the Stouffer method.}
}
\value{
Returns a list of master regulators, or a list of lists if a clustring is specified.
}
\description{
Returns the master regulators for the given data.
}
