% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-norm_funcs.R
\name{GetPCAFeats}
\alias{GetPCAFeats}
\title{Gets the number of PCA features to use from a Seurat object based on the given variance theshold.}
\usage{
GetPCAFeats(seurat.obj, var.thresh = 0.9)
}
\arguments{
\item{seurat.obj}{Seurat object with PCA reduction.}

\item{var.thresh}{Amount of variance to include. Default of 0.9.}
}
\value{
Number of PCA features to use.
}
\description{
Gets the number of PCA features to use from a Seurat object based on the given variance theshold.
}
