% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{group_colors}
\alias{group_colors}
\title{Identifies a vector of colors for a given number of groups.}
\usage{
group_colors(k, offset = 0)
}
\arguments{
\item{k}{Number of groups.}

\item{offset}{Optional argument to shift colors along color wheel.}
}
\value{
A vector of hues
}
\description{
Identifies a vector of colors for a given number of groups.
}
