% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{HasPISCESAssay}
\alias{HasPISCESAssay}
\title{Checks if the PISCES assay is present in a given object.}
\usage{
HasPISCESAssay(data.object)
}
\arguments{
\item{data.object}{Seurat object.}
}
\value{
Boolean TRUE or FALSE.
}
\description{
Checks if the PISCES assay is present in a given object.
}
