% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-regulators-data.R
\docType{data}
\name{human_regulators}
\alias{human_regulators}
\title{Human Regulator sets as Gene Names}
\format{
A list with four elements; each a list for a class of regulators.
}
\source{
Curated by the Califano lab. See ARACNe3 for more details.
}
\usage{
data(human_regulators)
}
\description{
Human regulator sets generated by the Califano lab for use with ARACNe3 and NaRnEA.
Regulators are separated into TFs, COTFs, Singaling, and Surface proteins.
Stored as a list of lists.
}
\examples{
data(human_regulators)
\donttest{length(human_regulators)}
}
\keyword{datasets}
