% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{internal_ges}
\alias{internal_ges}
\title{Generates a single-sample, internal GES.}
\usage{
internal_ges(
  filt.counts,
  norm.method = c("cpm", "pflpf"),
  est.method = c("map", "ps"),
  map.iter = 10
)
}
\arguments{
\item{filt.counts}{Matrix (features X samples) of filtered count data.}

\item{norm.method}{Normalization method. One of `c('cpm', 'pflpf')`, uses `pflpf` by default.}

\item{est.method}{Estimation method. One of `c('mle', 'map')`, with `map` by default.}

\item{map.iter}{Number of iterations to use if using MAP estimation. Default of 10.}
}
\value{
GES Matrix (features X samples).
}
\description{
Generates a single-sample, internal GES.
}
