% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{kappa_cluster_mrs}
\alias{kappa_cluster_mrs}
\title{Identifies cluster master regulators based on Cohen's Kappa.}
\usage{
kappa_cluster_mrs(narnea.res, clust.vec, num.mrs = 10)
}
\arguments{
\item{narnea.res}{NaRnEA result object with `PES` and `NES` members.}

\item{clust.vec}{Vector of clustering labels.}

\item{num.mrs}{Number of MRs to return.}
}
\value{
A list; table of positive MRs (`positive`) and negative MRs (`negative`); 
Stouffer integrated NaRnEA results by cluster (`cluster.narnea`); Kappa values for each protein in each cluster (`kappa.df`).
}
\description{
Identifies cluster master regulators based on Cohen's Kappa.
}
