% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne_funcs.R
\name{knn}
\alias{knn}
\title{Generates a KNN matrix.}
\usage{
knn(dist.mat, k = 5)
}
\arguments{
\item{dist.mat}{Distance matrix.}

\item{k}{Number of neighbors. Default of 5.}
}
\value{
A naighbor matrix with samples in rows and neighbors in columns.
}
\description{
Generates a KNN matrix.
}
