% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{kw_cluster_mrs}
\alias{kw_cluster_mrs}
\title{Identifies cluster-specific master regulators using a Kruskal-Wallis test.}
\usage{
kw_cluster_mrs(narnea.res, clust.vec, p.thresh = 0.05)
}
\arguments{
\item{narnea.res}{NaRnEA result object with `PES` and `NES` members.}

\item{clust.vec}{Vector of categorical labels.}

\item{p.thresh}{p-value threshold for calling significant master regulators. Default of 0.05.}
}
\value{
A list; table of positive MRs (`positive`) and negative MRs (`negative`); 
Kruskal-Wallis results (`kw.test`) with log p-values and group residuals..
}
\description{
Identifies cluster-specific master regulators using a Kruskal-Wallis test.
}
