% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{log_kw}
\alias{log_kw}
\title{Returns a log p-value from a Kruskal-Wallis test for the given data.}
\usage{
log_kw(feat.vec, group.vec)
}
\arguments{
\item{feat.vec}{Vector of values to be tested.}

\item{group.vec}{Vector of group labels. Assumed to be in the same order as `feat.vec`.}
}
\value{
A list; `log.p` containing the log p-value and `group.residuals` (group mean rank - overall mean rank).
}
\description{
Returns a log p-value from a Kruskal-Wallis test for the given data.
}
