% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{LogTTest}
\alias{LogTTest}
\title{Performs a t-test between two sample vectors x and y. Returns a log p-value.}
\usage{
LogTTest(x, y)
}
\arguments{
\item{x}{Vector of test values.}

\item{y}{Vector of reference values.}

\item{bootstrap.num}{Number of bootstraps to use. Default of 100.}
}
\value{
A signed log p-value.
}
\description{
Performs a t-test between two sample vectors x and y. Returns a log p-value.
}
