% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{louvain_clust}
\alias{louvain_clust}
\title{Runs Louvain clusering on the given object with the specified KNN graph.}
\usage{
louvain_clust(dist.mat, num.neighbors = 5)
}
\arguments{
\item{num.neighbors}{Number of neighbors to use in the KNN graph. Default of 5.}

\item{data.obj}{Seurat object w/ dist.mat object in misc of active assay.}
}
\value{
Vector of cluster assignments.
}
\description{
Runs Louvain clusering on the given object with the specified KNN graph.
}
