% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{louvain_k}
\alias{louvain_k}
\title{Optimizes Louvain clustering over a range of number of neighbors 'k' by silhouette score.}
\usage{
louvain_k(dist.mat, kmin = 5, kmax = 50, kstep = 5)
}
\arguments{
\item{dist.mat}{Distance matrix or distance structure like that generated by 'dist'.}

\item{kmin}{Minimum number of neighbors Default of 5.}

\item{kmax}{Maximum number of neibhbors Default of 50.}

\item{kstep}{Step size between k values. Default of 5.}
}
\value{
A list; 
'opt.clust' with the optimal clustering; 
'clustering.objs', a list of all clusterings generated; 
'sil.scores' a list of silhouette scores for each clustering object.
}
\description{
Optimizes Louvain clustering over a range of number of neighbors 'k' by silhouette score.
}
