% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{LouvainKRange}
\alias{LouvainKRange}
\title{Runs Louvain clustering w/ each of the specified number of neighbors used in the KNN graph.}
\usage{
LouvainKRange(data.obj, kmin = 5, kmax = 50, kstep = 5)
}
\arguments{
\item{data.obj}{Seurat object w/ dist.mat object in misc. of active assay.}

\item{kmin}{Minimum number of neighbors to use. Default of 5.}

\item{kmax}{Maximum number of neighbors to use. Default of 50.}

\item{kstep}{Steps to take between values of k. Default of 5.}
}
\value{
Updated data.obj w/ clustering.obj and pisces.cluster added to active assay.
}
\description{
Runs Louvain clustering w/ each of the specified number of neighbors used in the KNN graph.
}
