% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{LouvainResRange}
\alias{LouvainResRange}
\title{Louvain clustering over a range of resolution parameters (uses getComMembership from the MUDAN package)}
\usage{
LouvainResRange(data.obj, rmin = 10, rmax = 100, rstep = 10, verbose = TRUE)
}
\arguments{
\item{rmin}{Lowest resolution to try. Default of 10.}

\item{rmax}{Maximum resolution to try. Default of 100.}

\item{rstep}{Step size for resolution parameter. Default of 10.}

\item{verbose}{Switch to control terminal read out of progress. Default of TRUE.}

\item{data.object}{Seurat object w/ scale.data and a distance matrix in misc$dist.mat included.}
}
\description{
Louvain clustering over a range of resolution parameters (uses getComMembership from the MUDAN package)
}
