% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_funcs.R
\name{make_mds}
\alias{make_mds}
\title{Generates an MDS using base R functionality.}
\usage{
make_mds(dist.mat)
}
\arguments{
\item{dist.mat}{Distance matrix.}
}
\value{
Two column MDS matrix.
}
\description{
Generates an MDS using base R functionality.
}
