% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aracne_funcs.R
\name{make_metacells}
\alias{make_metacells}
\title{Generates meta cells for each cluster}
\usage{
make_metacells(
  counts.mat,
  dist.mat,
  clust.vec,
  num.neighbors = 5,
  subset = 250,
  min.samps = 500
)
}
\arguments{
\item{counts.mat}{Matrix of filtered counts (genes X samples).}

\item{dist.mat}{Distance matrix for this data.}

\item{clust.vec}{Clustering vector. If not specified, entire matrix will be used.}

\item{num.neighbors}{Number of neighbors to use in metacells. Default of 5.}

\item{subset}{Number of cells to subset to. Default of 250. No subsetting if set equal to NULL.}

\item{min.samps}{Minimum number of samples in a cluster required for meta cells. Default of 500.}
}
\value{
A list of meta cell matrices for all clusters with enough samples.
}
\description{
Generates meta cells for each cluster
}
