% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_funcs.R
\name{make_umap}
\alias{make_umap}
\title{Generates a UMAP using the uwot package.}
\usage{
make_umap(dat.mat, dense = TRUE)
}
\arguments{
\item{dat.mat}{Matrix of data (features X samples).}

\item{dense.map}{Flag to use dense UMAP. Default of TRUE.}
}
\value{
Two column matrix with UMAP representation of samples.
}
\description{
Generates a UMAP using the uwot package.
}
