% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-reduc_funcs.R
\name{MakeMDS}
\alias{MakeMDS}
\title{Builds a UMAP using the uwot package. Optionally can subset to most significant data within each sample.}
\usage{
MakeMDS(dat.object, num.dims = 2)
}
\arguments{
\item{dat.object}{Seurat object w/ PISCES assay and viper matrix in 'misc'.}

\item{num.dims}{Number of dimensions to return. Default of 2.}
}
\value{
Modified Seurat object w/ UMAP object in misc$umap
}
\description{
Builds a UMAP using the uwot package. Optionally can subset to most significant data within each sample.
}
