% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim-reduc_funcs.R
\name{MakeUMAP}
\alias{MakeUMAP}
\title{Builds a UMAP using the uwot package. Optionally can subset to most significant data within each sample.}
\usage{
MakeUMAP(dat.object, cbc = FALSE, num.feats = 10)
}
\arguments{
\item{dat.object}{Seurat object w/ PISCES assay and viper matrix in 'misc'.}

\item{cbc}{Flag to perform cell-by-cell UMAP. Default of FALSE.}

\item{num.feats}{Number of features to use if cbc flag is TRUE. Default of 10.}
}
\value{
A 2-column mantrix w/ UMAP coordinates.
}
\description{
Builds a UMAP using the uwot package. Optionally can subset to most significant data within each sample.
}
