% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{MarkerGrid}
\alias{MarkerGrid}
\title{Creates a grid of plots for the given markers that exist in the given matrix.}
\usage{
MarkerGrid(umap, cluster, pAct, markers, plotTitle)
}
\arguments{
\item{umap}{UMAP object generated from the 'UMAP' package}

\item{cluster}{Vector of cluster labels for each sample}

\item{pAct}{Matrix of protein activity (features X samples).}

\item{markers}{List of markers of interest.}

\item{plotTitle}{Title for the plot.}
}
\description{
Creates a grid of plots for the given markers that exist in the given matrix.
}
