% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_funcs.R
\name{matrix_convert_genes}
\alias{matrix_convert_genes}
\title{Converts the gene rownames of a matrix from the `from` to `to` convention.
Removes rows of the matrix for which no name in the `to` convention can be found.}
\usage{
matrix_convert_genes(dat.mat, species, from, to)
}
\arguments{
\item{dat.mat}{Matrix with rownames to be changed.}

\item{species}{Species for this data. Either 'human' or 'mouse'.}

\item{from}{Starting name convention. One of 'ensg', 'gn', or 'entrez'.}

\item{to}{Ending name convention. One of 'ensg', 'gn', or 'entrez'.}
}
\value{
Matrix with converted row names.
}
\description{
Converts the gene rownames of a matrix from the `from` to `to` convention.
Removes rows of the matrix for which no name in the `to` convention can be found.
}
