% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{meta_narnea}
\alias{meta_narnea}
\title{Runs MetaNaRnEA using the given ges matrix and network list.}
\usage{
meta_narnea(ges.mat, net.list, sample.weights = TRUE)
}
\arguments{
\item{ges.mat}{GES matrix (features X samples).}

\item{net.list}{List of A3 networks.}

\item{sample.weights}{Flag to perform weighted integration. Default of TRUE.
WARNING: Running w/ sample.weights = TRUE may be very slow for large numbers of samples.}
}
\value{
NaRnEA result list; matrices (regulators X samples) for NES and PES values.
}
\description{
Runs MetaNaRnEA using the given ges matrix and network list.
}
