% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwkmeans_funcs.R
\name{MinAssignments}
\alias{MinAssignments}
\title{Calculates the Phi matrix that minimizes the objective for each data.}
\usage{
MinAssignments(centers, data)
}
\arguments{
\item{centers}{Current cluster cents}

\item{data}{Data matrix (features X samples)}
}
\description{
Calculates the Phi matrix that minimizes the objective for each data.
}
