% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwkmeans_funcs.R
\name{MinClusters}
\alias{MinClusters}
\title{Calculates the optimal centers given data and cluster assignments.}
\usage{
MinClusters(phi, data)
}
\arguments{
\item{phi}{Matrix of phi values for each data.}

\item{data}{Data matrix (features X samples).}
}
\value{
Optimal center matrix
}
\description{
Calculates the optimal centers given data and cluster assignments.
}
