% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse-regulators-data.R
\docType{data}
\name{mouse_regulators}
\alias{mouse_regulators}
\title{Mouse Regulator sets as Gene Names}
\format{
A list with three elements; each a list for a class of regulators.
}
\source{
Curated by the Califano lab. See ARACNe3 for more details.
}
\usage{
data(mouse_regulators)
}
\description{
Mouse regulator sets generated by the Califano lab for use with ARACNe3 and NaRnEA.
Regulators are separated into TFs, COTFs, and Surface proteins.
Stored as a list of lists.
}
\examples{
data(mouse_regulators)
\donttest{length(mouse_regulators)}
}
\keyword{datasets}
