% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{MR_UnWrap}
\alias{MR_UnWrap}
\title{Unwraps a nested MR list: previous functions return cluster specific master regulators as a list of lists. This funciton will unwrap that object into one, unique list.}
\usage{
MR_UnWrap(MRs, top)
}
\arguments{
\item{MRs}{List of lists, with MR names as sub-list names and MR activity as sub-list entries.}

\item{top}{If specified, will subset the top X regulators from each set.}
}
\value{
Returns a de-duplicated list of MRs.
}
\description{
Unwraps a nested MR list: previous functions return cluster specific master regulators as a list of lists. This funciton will unwrap that object into one, unique list.
}
