% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{MRHeatmap}
\alias{MRHeatmap}
\title{Generates a heatmap of the cluster-specific master regulators.}
\usage{
MRHeatmap(pisces.obj, num.mrs = 10, clust.vect, plot.title = "")
}
\arguments{
\item{pisces.obj}{Seurat object w/ PISCES assay, VIPER matrix, and master regulators.}

\item{num.mrs}{Number of master regulators to include for each plot. Default of 10.}

\item{clust.vect}{Optional argument for cluster labels; if not included, uses "PISCES;misc$pisces.cluster"}

\item{plot.title}{Optional plot title argument. If not specified, no title included.}
}
\value{
Returns the pheatmap plot object.
}
\description{
Generates a heatmap of the cluster-specific master regulators.
}
