% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{MRTableWrite}
\alias{MRTableWrite}
\title{Writes a given master regulator object to tsvs for each group.}
\usage{
MRTableWrite(
  mr.obj,
  file.dir,
  file.name,
  num.mrs = 50,
  top = TRUE,
  bottom = FALSE
)
}
\arguments{
\item{mr.obj}{List of master regulators; list of lists, w/ positive and negative lists of MRs for each group}

\item{file.dir}{Directory to write file.}

\item{file.name}{Name to be used for each table.}

\item{num.mrs}{Number of MRs to write. Default of 50.}

\item{top}{Optional argument to write top (activated) MRs. TRUE by default.}

\item{bottom}{Optional argument to write bottom (deactivated) MRs. FALSE by default.}
}
\description{
Writes a given master regulator object to tsvs for each group.
}
