% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt-genes-data.R
\docType{data}
\name{mt_genes}
\alias{mt_genes}
\title{Mosue / human mitochondrial genes in either GN or ENSG format.}
\format{
A dataframe with four columns; one for each of hum / mur + ensg / symb.
}
\source{
Curated by the Califano lab.
}
\usage{
data(mt_genes)
}
\description{
Mosue / human mitochondrial genes in either GN or ENSG format.
}
\examples{
data(mt_genes)
\donttest{length(mt_genes)}
}
\keyword{datasets}
