% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_funcs.R
\name{MTFilter}
\alias{MTFilter}
\title{Filters data based on percentage of mitochondrial gens.}
\usage{
MTFilter(dat.mat, mt.genes, mt.thresh = 0.1)
}
\arguments{
\item{mt.genes}{Path to .csv file with ENSG and Hugo names for mitochondrial genes.}

\item{mt.thresh}{Threshold above which cells will be removed. Default of 0.15}

\item{raw.mat}{Matrix of raw gene expression data (genes X samples)}
}
\description{
Filters data based on percentage of mitochondrial gens.
}
