% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwkmeans_funcs.R
\name{MWKMeans}
\alias{MWKMeans}
\title{Generates a multiway k-means given the data and starting center points.}
\usage{
MWKMeans(data, centers)
}
\arguments{
\item{data}{Matrix of data (features X samples)}

\item{centers}{Starting centers in matrix format (features X centers)}
}
\value{
A list; 'phi' with the cluster assignments for each data and 'centers' with the calculated cluster centers.
}
\description{
Generates a multiway k-means given the data and starting center points.
}
