% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{narnea_combine}
\alias{narnea_combine}
\title{Generates two 3D matrices (PES/NES) from a list of NaRnEA results.}
\usage{
narnea_combine(narnea.list)
}
\arguments{
\item{narnea.list}{List of result objects from `MatrixNaRnEA`}
}
\value{
Two member list - PES and NES - with each element a 3D matrix.
}
\description{
Generates two 3D matrices (PES/NES) from a list of NaRnEA results.
}
