% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{narnea_dist}
\alias{narnea_dist}
\title{Generates a distance matrix from NaRnEA results sample-sample PES enrichment.}
\usage{
narnea_dist(pes.mat, num.prots = 30)
}
\arguments{
\item{pes.mat}{Matrix of PES values (proteins X samples).}

\item{num.prots}{Number of proteins to use in each regulon. Default/minimum of 30. 
WARNING: Using more proteins will increase runtime.}
}
\value{
A distance matrix (samples X samples).
}
\description{
Generates a distance matrix from NaRnEA results sample-sample PES enrichment.
}
