% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{nes_covariance}
\alias{nes_covariance}
\title{Compute the Covariance of the Directed and Undirected NES.}
\usage{
nes_covariance(R, S, AW.AM.prod, AW.AM.abs.prod, E.r, E.rs, D.v, U.v, g)
}
\arguments{
\item{R}{Rank normalized ges matrix (genes x samples).}

\item{S}{Sign of ges matrix (genes x samples).}

\item{AW.AM.prod}{Product of the association weight (AW) and mode (AM) matrices created from the regulon list.}

\item{AW.AM.abs.prod}{Product of the association weight (AW) and absolute value of the association mode (AM.abs) matrices created from the regulon list.}

\item{E.r}{Expected value of r in each sample (scalar value).}

\item{E.rs}{Expected value of r*s in each sample (1 x samples).}

\item{D.v}{Variance of Directed Enrichment Score (regulons x samples).}

\item{U.v}{Variance of Undirected Enrichment Score (regulons x samples).}

\item{g}{Number of genes in the gene expression signature matrix (scalar value).}
}
\value{
Matrix of covariance values (regulons x samples).
}
\description{
Compute the Covariance of the Directed and Undirected NES.
}
