% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{network_match_narnea}
\alias{network_match_narnea}
\title{Performs NaRnEA on multiple samples using network matching. Those samples with matched networks
will have protein activity generated with one network. Those without will be analyzed using MetaNaRnEA.}
\usage{
network_match_narnea(ges.mat, net.list, net.match.vec)
}
\arguments{
\item{ges.mat}{GES matrix (genes X samples).}

\item{net.list}{List of networks appropriately parametrized for NaRnEA.}

\item{net.match.vec}{Named vector of the same length as the number of samples, with either the name of a matched network or NA}
}
\value{
NaRnEA object
}
\description{
Performs NaRnEA on multiple samples using network matching. Those samples with matched networks
will have protein activity generated with one network. Those without will be analyzed using MetaNaRnEA.
}
