% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{NormalizeExpression}
\alias{NormalizeExpression}
\title{Normalizes gene expression in the PISCESgexp assay. Stores normalized counts in 'data' field.
Additionally, generates a scaled matrix (interpretable as an internal GES) to the 'scale.data' field.}
\usage{
NormalizeExpression(pisces.obj, method = c("cpm", "pflogpf"), filt.per = 0.01)
}
\arguments{
\item{pisces.obj}{A Seurat object with the PISCESgexp assay.}

\item{method}{One of 'cpm' or 'pflogpf'}

\item{filt.per}{Threshold value for gene filtration; genes w/ less than filt.per * num.samps total counts will be removed. Default of 0.01.}
}
\value{
A Seurat object with normalized expression in the 'data' field of the PISCESgexp assay.
}
\description{
Normalizes gene expression in the PISCESgexp assay. Stores normalized counts in 'data' field.
Additionally, generates a scaled matrix (interpretable as an internal GES) to the 'scale.data' field.
}
