% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwkmeans_funcs.R
\name{ObjectiveValue}
\alias{ObjectiveValue}
\title{Calculates the value of the objective function for optimizing cluster assignments.}
\usage{
ObjectiveValue(j, jPrime, x)
}
\arguments{
\item{j}{}
}
\description{
Calculates the value of the objective function for optimizing cluster assignments.
}
