% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{opt_kappa}
\alias{opt_kappa}
\title{Determines the optimal possible Cohen's Kappa that can be generated by the given feature.}
\usage{
opt_kappa(class.vec, feature.vec)
}
\arguments{
\item{class.vec}{Vector of categorical labels.}

\item{feature.vec}{Vector of numerical feature values. Assumes the same ordering as `class.vec`.}
}
\value{
THe optimal possible value of Cohen's Kappa for a classifier generated using this feature.
}
\description{
Determines the optimal possible Cohen's Kappa that can be generated by the given feature.
}
