% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{pairwise_cluster_mrs}
\alias{pairwise_cluster_mrs}
\title{Identified MRs based on a pairwise comparison of clusters.}
\usage{
pairwise_cluster_mrs(norm.counts, clust.vec, net.list, num.mrs = 10)
}
\arguments{
\item{norm.counts}{Matrix of normalized counts (features X samples).}

\item{clust.vec}{Clustering vector.}

\item{net.list}{List of ARACNe3 networks.}

\item{num.mrs}{Number of master regulators to return for each cluster. Default of 10.}
}
\value{
A three member list w/ `positive` and `negative` MRs as well as the full NaRnEA results `mr.narnea`.
}
\description{
Identified MRs based on a pairwise comparison of clusters.
}
