% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{pam_k}
\alias{pam_k}
\title{Optimizes PAM clustering over a range of cluter numbers 'k' using silhouette score.}
\usage{
pam_k(dist.mat, kmin = 2, kmax = 5)
}
\arguments{
\item{dist.mat}{Distance matrix or distance structure like that generated by 'dist'.}

\item{kmin}{Minimum number of clusters. Default of 2.}

\item{kmax}{Maximum number of clusters. Default of 5.}
}
\value{
A list; 
'opt.clust' with the optimal clustering; 
'clustering.objs', a list of all clusterings generated; 
'sil.scores' a list of silhouette scores for each clustering object.
}
\description{
Optimizes PAM clustering over a range of cluter numbers 'k' using silhouette score.
}
