% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{PAMClust}
\alias{PAMClust}
\title{Performs PAM clustering on the active assay of the given Seurat object.
Looks for a distance matrix 'dist' in the 'misc' field of the active assay.}
\usage{
PAMClust(pisces.obj, kmin = 2, kmax = 5)
}
\arguments{
\item{pisces.obj}{A Seurat object w/ either 'dist' in the 'misc' field or w/ 'scale.data'.}

\item{kmin}{Minimum number of clusters. Default of 2.}

\item{kmax}{Maximum number of clusters. Default of 5.}
}
\value{
Seurat object with clustering object 'louvain.clust' added to 'misc' field of active assay.
And optimal clustering set as 'pisces.cluster' label for active assay.
}
\description{
Performs PAM clustering on the active assay of the given Seurat object.
Looks for a distance matrix 'dist' in the 'misc' field of the active assay.
}
