% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{PamKRange}
\alias{PamKRange}
\title{PAM clustering over a range of k values.}
\usage{
PamKRange(dist.mat, kmin = 2, kmax = 5, verbose = TRUE)
}
\arguments{
\item{dist.mat}{A distance matrix for the data to be clustered.}

\item{kmin}{The minimum k to be used (default of 2).}

\item{kmax}{The maximum k to be used (default of 5).}

\item{verbose}{Switch to print status updates.}
}
\value{
A list of two lists; 'clusterings', which contains the cluster labels and 'sils', which has silhouette scores.
}
\description{
PAM clustering over a range of k values.
}
