% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc-raw-data.R
\docType{data}
\name{pbmc.raw}
\alias{pbmc.raw}
\title{Single-Cell RNASeq from PBMC data}
\format{
A matrix.
}
\source{
"[insert citation]"
}
\usage{
data(pbmc.raw)
}
\description{
Human peripheral blood mononuclear cells sequenced using the 10x Chromium platform.
This dataset has been subset to 1,000 samples and contains 58,828 genes. Genes are in rows
and samples are in columns. Genes are labeled with HGNC names.
}
\examples{
data(pbmc.raw)
\donttest{dim(pbmc.raw)}
}
\keyword{datasets}
