% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{pf_norm}
\alias{pf_norm}
\title{Normalizes a matrix using one round of proportional fitting.}
\usage{
pf_norm(dat.mat, sf)
}
\arguments{
\item{dat.mat}{Matrix of data (features X samples).}

\item{sf}{Optional size factor argument. If not specified, will be computed as mean of column sums.}
}
\value{
Normalized matrix (features X samples).
}
\description{
Normalizes a matrix using one round of proportional fitting.
}
