% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{pflpf_norm}
\alias{pflpf_norm}
\title{Performs PF log1 PF norm normalization.}
\usage{
pflpf_norm(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of data (features X samples).}
}
\value{
Normalized matrix (features X samples).
}
\description{
Performs PF log1 PF norm normalization.
}
