% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{PISCESActivity}
\alias{PISCESActivity}
\title{Generates protein activity readout using the NaRnEA algorithm.
Uses either the GES object found in the `misc` field of the `PISCESpact` assay or
the `scale.data` field of the `PISCESgexp` assay.}
\usage{
PISCESActivity(pisces.obj, net.list, sample.weights = TRUE, use.sct = FALSE)
}
\arguments{
\item{pisces.obj}{Seurat object w/ the `PISCESpact` assay.}

\item{net.list}{List of networks. If not specified, networks in the `misc` field of the `PISCESpact` assay will be used.}

\item{sample.weights}{Flag to compute sample-specific weights. Default of TRUE.}

\item{use.sct}{Flag to use the SCT matrix as a GES. Default of FALSE.}
}
\value{
Seurat object with `PES` matrix in `scale.data` of `PISCESpact` assay.
`NES` results will be stored in the `misc` field of the `PISCESpact` assay.
}
\description{
Generates protein activity readout using the NaRnEA algorithm.
Uses either the GES object found in the `misc` field of the `PISCESpact` assay or
the `scale.data` field of the `PISCESgexp` assay.
}
