% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{PISCESViper}
\alias{PISCESViper}
\title{Runs VIPER on a seurat.obj that has the appropriately generated 'PISCES' assay and scaled data.}
\usage{
PISCESViper(data.object, net.list, sct.ges = FALSE)
}
\arguments{
\item{data.object}{Seurat object w/ 'PISCES' assay and scaled data.}

\item{net.list}{List of networks OR a single network.}

\item{sct.ges}{Optional flag to use SCT scale data as GES. Default of False.}
}
\value{
Seurat.object with added 'viper' matrix in PISCES assay
}
\description{
Runs VIPER on a seurat.obj that has the appropriately generated 'PISCES' assay and scaled data.
}
