% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{qc_filt}
\alias{qc_filt}
\title{Filters raw gene expression data based on depth, unique genes, and percentage of MT reads.}
\usage{
qc_filt(
  raw.counts,
  min.depth = 0,
  max.depth = Inf,
  min.genes = 0,
  max.genes = Inf,
  max.mt = 1,
  min.gene.count,
  species = c("hum", "mur"),
  genes = c("symb", "ensg")
)
}
\arguments{
\item{raw.counts}{Matrix of raw gene expression data (features X samples).}

\item{min.depth}{Minimum depth for each sample. Default of 0.}

\item{max.mt}{Maximum percentage of Mitochondrial reads. Default of 1.}

\item{min.gene.count}{Minimum total reads for a gene to be kept. If not specified, default is 0.01 * # of samples.}

\item{species}{One of `c('hum', 'mur')`, specifying human or murine data respectively. If not specified, assumes human.}

\item{genes}{One of `c('symb', 'ensg')`, specifying gene symbols or ENSG names respectively. If not specified, assumed symbols.}
}
\value{
Matrix of normalized counts (features X samples).
}
\description{
Filters raw gene expression data based on depth, unique genes, and percentage of MT reads.
}
