% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{qc_plot}
\alias{qc_plot}
\title{Generates QC violin plots for depth, detected genes, and MT% metrics.}
\usage{
qc_plot(raw.counts, species = c("hum", "mur"), genes = c("symb", "ensg"))
}
\arguments{
\item{raw.counts}{Matrix of raw gene expression data (features X samples).}

\item{species}{One of `c('hum', 'mur')`, specifying human or murine data respectively. If not specified, assumes human.}

\item{genes}{One of `c('symb', 'ensg')`, specifying gene symbols or ENSG names respectively. If not specified, assumed symbols.}
}
\value{
A ggplot object.
}
\description{
Generates QC violin plots for depth, detected genes, and MT% metrics.
}
