% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-norm_funcs.R
\name{QCPlots}
\alias{QCPlots}
\title{Generates basic quality control plots from raw gene expression data.}
\usage{
QCPlots(raw.mat, mt.genes, plot.path)
}
\arguments{
\item{raw.mat}{Matrix of raw gene expression data (genes X samples).}

\item{mt.genes}{Path to .csv file with ENSG and Hugo names for mitochondrial genes.}

\item{plot.path}{Optional argumetn of save path for plot.}
}
\description{
Generates basic quality control plots from raw gene expression data.
}
