% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-norm_funcs.R
\name{QCTransform}
\alias{QCTransform}
\title{Filters out genes with no expression and low quality cells.}
\usage{
QCTransform(raw.mat, minCount = 1000, maxCount = 1e+05, minGeneReads = 1)
}
\arguments{
\item{raw.mat}{Matrix of raw gene expression data (genes X samples).}

\item{minCount}{Minimum number of reads in a cell. Default of 1000.}

\item{maxCount}{Maximum number of reads in a cell. Default of 100000.}

\item{minGeneReads}{Minimum number of reads for a gene to be kept. Default of 1 (any gene with no reads will be removed).}
}
\value{
Quality controlled matrix.
}
\description{
Filters out genes with no expression and low quality cells.
}
