% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{quantile_breaks}
\alias{quantile_breaks}
\title{Generates breaks for a color scale based on quantiles.}
\usage{
quantile_breaks(dat.mat, n)
}
\arguments{
\item{dat.mat}{Data matrix (features X samples).}

\item{n}{Number of breaks to generate. If not specified, uses first three stdevs.}
}
\value{
Numeric vector of break values.
}
\description{
Generates breaks for a color scale based on quantiles.
}
