% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{QuantileBreaks}
\alias{QuantileBreaks}
\title{Generates breaks for a color scale based on quantiles.}
\usage{
QuantileBreaks(xs, n)
}
\arguments{
\item{n}{Number of breaks to generate. If not specified, uses first three stdevs.}

\item{dat.mat}{Data matrix (features X samples).}
}
\value{
Numeric vector of break values.
}
\description{
Generates breaks for a color scale based on quantiles.
}
