% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-norm_funcs.R
\name{RankTransform}
\alias{RankTransform}
\title{Performs a rank transformation on a given matrix.}
\usage{
RankTransform(dat.mat)
}
\arguments{
\item{dat.mat}{Matrix of data, usually gene expression (genes X samples).}
}
\value{
Rank transformed matrix.
}
\description{
Performs a rank transformation on a given matrix.
}
