% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_funcs.R
\name{read10X}
\alias{read10X}
\title{Read in data in 10x format}
\usage{
read10X(dat.path)
}
\arguments{
\item{dat.path}{Path to directory containing 'matrix.mtx', 'genes.tsv', and 'barcodes.tsv'.}
}
\value{
Matrix of raw gene expression (genes X samples).
}
\description{
Read in data in 10x format
}
