% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp-transform_funcs.R
\name{scale_ges}
\alias{scale_ges}
\title{Generates a gene expression signature by scaling and centering, either againt a reference or internally.}
\usage{
scale_ges(dat.mat, ref.mat)
}
\arguments{
\item{dat.mat}{Matrix of test data (features X samples).}

\item{ref.mat}{Optional matrix of reference data (features X samples). If not specified, an internal GES will be generated.}
}
\value{
GES Matrix (featues X samples).
}
\description{
Generates a gene expression signature by scaling and centering, either againt a reference or internally.
}
