% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{sig_features}
\alias{sig_features}
\title{Returns a list of significant proteins for each sample (column) in the given NES matrix.
Identifies significant proteins by Benjamini-Hochberg at the given significance level.}
\usage{
sig_features(nes.mat, p.val.thresh = 0.05)
}
\arguments{
\item{nes.mat}{Matrix of NES scores (features X columns), such as that generated by NaRnEA analysis.}

\item{p.val.thresh}{p-value threshold. Default of 0.05.}
}
\value{
List of lists; one list for each sample (column), containing the names of significant proteins.
}
\description{
Returns a list of significant proteins for each sample (column) in the given NES matrix.
Identifies significant proteins by Benjamini-Hochberg at the given significance level.
}
