% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{SilScoreEval}
\alias{SilScoreEval}
\title{Generation of silhouette score for a list of clusters.}
\usage{
SilScoreEval(clusterings, dist.mat, plotPath)
}
\arguments{
\item{clusterings}{List of clustering objects.}

\item{dist.mat}{Distance matrix for the data clustered in clusterings.}

\item{plotPath}{If specified, will save a plot of the silhouette scores.}
}
\value{
List of silhouette scores.
}
\description{
Generation of silhouette score for a list of clusters.
}
