% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_funcs.R
\name{SSLouvain}
\alias{SSLouvain}
\title{Silhouette score optimized Louvain clustering.}
\usage{
SSLouvain(seurat.obj, dist.mat, rstep = 0.05, rmax = 0.5)
}
\arguments{
\item{seurat.obj}{Seurat object with Neighbors detected.}

\item{dist.mat}{Distance matrix for use w/ Silhouette score.}

\item{rstep}{Size of the steps to take in resolution. Default of 0.05.}

\item{rmax}{Maximum resolution to try. Default of 0.5.}
}
\value{
Seurat object with all clustering results and optimal result as active.ident.
}
\description{
Silhouette score optimized Louvain clustering.
}
