% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{stouffer_cluster_mrs}
\alias{stouffer_cluster_mrs}
\title{Returns an MR object with MRs selected via Stouffer integration.}
\usage{
stouffer_cluster_mrs(narnea.res, clust.vec, p.thresh = 0.05)
}
\arguments{
\item{narnea.res}{NaRnEA result object with `PES` and `NES` members.}

\item{clust.vec}{Vector of categorical labels.}

\item{p.thresh}{p-value threshold for calling significant master regulators. Default of 0.05.}
}
\value{
A list; table of positive MRs (`positive`) and negative MRs (`negative`); 
Stouffer integrated NaRnEA results by cluster (`cluster.narnea`).
}
\description{
Returns an MR object with MRs selected via Stouffer integration.
}
