% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{stouffer_narnea}
\alias{stouffer_narnea}
\title{Returns cluster-integrated NaRnEA results using Stouffer's method.
PES values are averaged across groups, while NES values are Stouffer integrated.}
\usage{
stouffer_narnea(narnea.res, clust.vec)
}
\arguments{
\item{narnea.res}{NaRnEA result object with `PES` and `NES` members.}

\item{clust.vec}{Vector of categorical labels.}
}
\value{
Integrated NaRnEA results; list with both `PES` and `NES` members.
}
\description{
Returns cluster-integrated NaRnEA results using Stouffer's method.
PES values are averaged across groups, while NES values are Stouffer integrated.
}
