% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{StoufferIntegrate}
\alias{StoufferIntegrate}
\title{Stouffer integrates the given vector of data.}
\usage{
StoufferIntegrate(dat.vect, weight.vect)
}
\arguments{
\item{dat.vect}{Vector of data to be integrated}

\item{weight.vect}{Vector of weights, if specified.}
}
\value{
Stouffer integrated value for the given data.
}
\description{
Stouffer integrates the given vector of data.
}
