% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{StoufferMRs}
\alias{StoufferMRs}
\title{Identifies MRs for given data using stouffer integration.}
\usage{
StoufferMRs(dat.mat, cluster, weights)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{cluster}{Vector of cluster lables. If not included, integrates the entire matrix.}

\item{weights}{A named vector of sample weights. If included, stouffer integration is weighted.}
}
\value{
Returns the stouffer integrated scores for each protien.
}
\description{
Identifies MRs for given data using stouffer integration.
}
