% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{TwoMatVipSim}
\alias{TwoMatVipSim}
\title{Returns the viperSimilarity of the elements of two matrices.}
\usage{
TwoMatVipSim(mat.1, mat.2)
}
\arguments{
\item{mat.1}{First matrix (features X columns).}

\item{mat.2}{Second matrix (features X columns).}
}
\value{
A matrix w/ viperSimilarity for elements of mat.1 crossed with elements of mat.2.
}
\description{
Returns the viperSimilarity of the elements of two matrices.
}
