% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{undirected_nes}
\alias{undirected_nes}
\title{Computes the Undirected NES}
\usage{
undirected_nes(R, S, AW.AM.abs.prod, E.r, E.r2, n)
}
\arguments{
\item{R}{Rank normalized ges matrix (genes x samples).}

\item{S}{Sign of ges matrix (genes x samples).}

\item{AW.AM.abs.prod}{Product of the association weight (AW) and absolute value of the association mode (AM.abs) matrices created from the regulon list.}

\item{E.r}{Expected value of r in each sample (scalar value).}

\item{E.r2}{Expected value of r^2 in each sample (scalar value).}

\item{n}{Number of samples (scalar value).}
}
\value{
List of matrices, each (regulons x samples):
Enrichment score matrix 'es'; expected value 'exp'; variance 'var'; NES matrix 'NES'
}
\description{
Computes the Undirected NES
}
