% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narnea_funcs.R
\name{unweighted_integration}
\alias{unweighted_integration}
\title{Integrates 3D arrays of NaRnEA results generated by `narnea_combine` with equal weights across results.}
\usage{
unweighted_integration(combine.list)
}
\arguments{
\item{combine.list}{List generated by `narnea_combine`; two 3D arrays, one each for PES and NES}
}
\value{
Integrated NaRnEA results, with PES and NES matrices.
}
\description{
Integrates 3D arrays of NaRnEA results generated by `narnea_combine` with equal weights across results.
}
