% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{ViperMerge}
\alias{ViperMerge}
\title{Merges two viper matrices, giving priority to one over the other.}
\usage{
ViperMerge(p.mat, q.mat)
}
\arguments{
\item{p.mat}{Priority viper matrix (proteins X samples). Proteins here will override those in the other matrix.}

\item{q.mat}{Secondary viper matrix (proteins X samples). Proteins here will fin in for gaps in the priority matrix.}
}
\value{
A merged viper matrix.
}
\description{
Merges two viper matrices, giving priority to one over the other.
}
