% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{WeightedVIPER}
\alias{WeightedVIPER}
\title{MetaVIPER implementation that will perform a weighted stouffer integration based on highest NES.}
\usage{
WeightedVIPER(ges, net.list, use.nets, ret.weights = FALSE)
}
\arguments{
\item{ges}{Gene Expression Signature (features X samples)}

\item{net.list}{List object with the networks to be used}

\item{use.nets}{Optional argument to sslect the top n networks. If not specified, all networks are used.}

\item{ret.weights}{Optional argument to return the network weight matrix as well as the VIPER matrix. FALSE by default.}
}
\value{
Either a viper matrix, or a list with a viper matrix and the network weight matrix.
}
\description{
MetaVIPER implementation that will perform a weighted stouffer integration based on highest NES.
}
