% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_funcs.R
\name{WriteMetaCells}
\alias{WriteMetaCells}
\title{Generates metacells from the given Seurat object. Rquires that the 'PISCESgexp' assay has been added.
By default, uses the 'pisces.clust' metadata to group clusters. This can be overrided with the 'clust.vec' argument.
If clust.vec is 'NA', then the entire matrix will be used.}
\usage{
WriteMetaCells(
  pisces.obj,
  file.path = ".",
  proj.name = "",
  clust.vec = NULL,
  num.neighbors = 5,
  subset = 250,
  min.samps = 500
)
}
\arguments{
\item{pisces.obj}{A Seurat object with the PISCESgexp assay.}

\item{proj.name}{File name preface}

\item{clust.vec}{Optional clustering vector object.}

\item{num.neighbors}{Number of neighbors to use in metacells. Default of 5.}

\item{subset}{Number of cells to subset to. Default of 250. No subsetting if set equal to NULL.}

\item{min.samps}{Minimum number of samples in a cluster required for meta cells. Default of 500.}

\item{out.dir}{Output directory for metacell matrices to be written to.}
}
\description{
Generates metacells from the given Seurat object. Rquires that the 'PISCESgexp' assay has been added.
By default, uses the 'pisces.clust' metadata to group clusters. This can be overrided with the 'clust.vec' argument.
If clust.vec is 'NA', then the entire matrix will be used.
}
