% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{AddPISCESAssay}
\alias{AddPISCESAssay}
\title{Adds a named assay 'PISCES' to a Seurat object.}
\usage{
AddPISCESAssay(seurat.obj)
}
\arguments{
\item{seurat.obj}{Seurat object w/ RNA assay.}
}
\value{
Seurat object w/ original assay and added 'PISCES' assay.
}
\description{
Adds a named assay 'PISCES' to a Seurat object.
}
