% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{cluster_mr_heatmap}
\alias{cluster_mr_heatmap}
\title{Generates a heatmap of cluster-specific MR activity.
Alternatively, can be provided with a pre-selected set of proteins, typically canonical markers.}
\usage{
cluster_mr_heatmap(
  dat.mat,
  dat.type = c("gexp", "pact"),
  clust.vec,
  mr.obj,
  num.mrs = 10,
  marker.set = NULL
)
}
\arguments{
\item{dat.mat}{Matrix of either scaled expresison data or PES generated by NaRnEA (featues X samples).}

\item{dat.type}{One of `c('gexp', 'pact')`, specifying expression or activity data. ASsumes activity if not specified.}

\item{clust.vec}{Named clustering vector.}

\item{mr.obj}{Master regulator object, such as that generated by `cluster_signature_mrs`.
Should be a list with a `cluster.narnea` element, containing cluster-specific NaRnEA results (NES and PES matrices).}

\item{num.mrs}{Number of master regulators to display. Default of 10.}

\item{marker.set}{Optional list of markers to use in lieu of master regulators.}
}
\value{
A ComplexHeatmap object.
}
\description{
Generates a heatmap of cluster-specific MR activity.
Alternatively, can be provided with a pre-selected set of proteins, typically canonical markers.
}
