% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{MarkerGrid}
\alias{MarkerGrid}
\title{Creates a grid of plots for the given markers that exist in the given matrix.}
\usage{
MarkerGrid(vip.obj, marker.list)
}
\arguments{
\item{marker.list}{List of gene names w/ markers of interest.}

\item{data.object}{Seurat object w/ PISCES assay, VIPER, and misc$umap}
}
\value{
Plot object from ggarrange
}
\description{
Creates a grid of plots for the given markers that exist in the given matrix.
}
