% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_funcs.R
\name{MTPercent}
\alias{MTPercent}
\title{Returns vector of mitochondrial percentages for the given samples.}
\usage{
MTPercent(dat.mat, mt.genes)
}
\arguments{
\item{dat.mat}{Matrix of raw gene expression data (genes X samples).}
}
\value{
Returns named vector of mitochondrial gene percentage.
}
\description{
Returns vector of mitochondrial percentages for the given samples.
}
