% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{MWUMrs}
\alias{MWUMrs}
\title{Identifies cluster specific master regulators using the Mann-Whitney U-test.
Approximates p-vals using a normal distribution for n > 30.}
\usage{
MWUMrs(dat.mat, clust.vect)
}
\arguments{
\item{dat.mat}{Matrix of data.}

\item{clust.vect}{Vector of cluster labels.}
}
\value{
For each cluster, two sorted lists of log p-values, split by positive / negative
}
\description{
Identifies cluster specific master regulators using the Mann-Whitney U-test.
Approximates p-vals using a normal distribution for n > 30.
}
