% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R
\name{MWUMrs}
\alias{MWUMrs}
\title{Identifies cluster specific master regulators using the Mann-Whitney U-test.
Approximates p-vals using a normal distribution for n > 30.}
\usage{
MWUMrs(dat.object, clust.vect)
}
\arguments{
\item{dat.object}{Seurat object w/ PISCES assay and viper matrix in 'misc'.}

\item{clust.vect}{Vector of cluster labels.}
}
\value{
List of log p-values for pos/neg MRs in each cluster; stored in PISCES assay under 'misc' as 'mwuMRs'
}
\description{
Identifies cluster specific master regulators using the Mann-Whitney U-test.
Approximates p-vals using a normal distribution for n > 30.
}
