% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper_funcs.R
\name{PISCESViper}
\alias{PISCESViper}
\title{Runs VIPER on a seurat.obj that has the appropriately generated 'PISCES' assay and scaled data.}
\usage{
PISCESViper(seurat.obj, net.list)
}
\arguments{
\item{seurat.obj}{Seurat object w/ 'PISCES' assay and scaled data.}

\item{net.list}{List of networks OR a single network.}
}
\value{
Seurat.object with added 'viper' matrix in PISCES assay
}
\description{
Runs VIPER on a seurat.obj that has the appropriately generated 'PISCES' assay and scaled data.
}
