% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{QCPlots}
\alias{QCPlots}
\title{Generates basic quality control plots from raw gene expression data.}
\usage{
QCPlots(seurat.obj)
}
\arguments{
\item{seurat.obj}{A Seurat object w/ MT% added to metadata.}
}
\description{
Generates basic quality control plots from raw gene expression data.
}
