% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{QCPlots}
\alias{QCPlots}
\title{Generates basic quality control plots from raw gene expression data.}
\usage{
QCPlots(seurat.obj, plot.path)
}
\arguments{
\item{seurat.obj}{A Seurat object w/ MT% added to metadata.}

\item{plot.path}{Optional argumetn of save path for plot.}
}
\description{
Generates basic quality control plots from raw gene expression data.
}
