% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{QuantileBreaks}
\alias{QuantileBreaks}
\title{Genreates breaks for a color scale based on quantiles.}
\usage{
QuantileBreaks(xs, n = 10)
}
\arguments{
\item{n}{Number of breaks to generate. Default of 10.}

\item{dat.mat}{Data matrix (features X samples).}
}
\value{
Numeric vector of break values.
}
\description{
Genreates breaks for a color scale based on quantiles.
}
