% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_funcs.R
\name{RankTransform}
\alias{RankTransform}
\title{Performs a rank transformation on a given matrix, typically as an alternative GES generation technique.
If a seurat object, stores results in scale.data field; otherwise, returns rank transformation  matrix.}
\usage{
RankTransform(data.object)
}
\arguments{
\item{data.object}{Either a Seurat object or a matrix of raw count data (genes X samples).}
}
\value{
Rank transformed matrix, or appropriately adjusted seurat object.
}
\description{
Performs a rank transformation on a given matrix, typically as an alternative GES generation technique.
If a seurat object, stores results in scale.data field; otherwise, returns rank transformation  matrix.
}
