% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_funcs.R, R/viper_funcs.R
\name{StoufferMRs}
\alias{StoufferMRs}
\title{Master regulators by Stouffer integration.}
\usage{
StoufferMRs(dat.mat, cluster, weights)

StoufferMRs(dat.mat, cluster, weights)
}
\arguments{
\item{dat.mat}{Matrix of protein activity (proteins X samples).}

\item{cluster}{Vector of cluster lables. If not included, integrates the entire matrix.}

\item{weights}{A named vector of sample weights. If included, stouffer integration is weighted.}

\item{clust.vect}{Vector of cluster labels.}

\item{weight.vect}{Vector of weights, if specified.}
}
\value{
Integrated

Returns the stouffer integrated scores for each protien.
}
\description{
Master regulators by Stouffer integration.

Identifies MRs for given data using stouffer integration.
}
